rename to issues to avoid common package import of 'errors'
