+++ internal/lexer/scanner/error.go
@@ -3,10 +3,10 @@ package scanner
 import (
 	"runtime"
 
-	"github.com/yoheimuta/go-protoparser/errors"
+	"github.com/yoheimuta/go-protoparser/issues"
 )
 
-func (s *Scanner) unexpected(found rune, expected string) errors.ParseError {
+func (s *Scanner) unexpected(found rune, expected string) issues.ParseError {
 	_, file, line, _ := runtime.Caller(1)
-	return errors.NewParseError(string(found), expected, file, line)
+	return issues.NewParseError(string(found), expected, file, line)
 }
